requirements/common.txt
pydantic-settings==2.0.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9

fastapi==0.103.2
uvicorn==0.23.2
Flask==3.0.0
